+++ releases/ChangeLog.md
@@ -2,6 +2,7 @@
 
 - Enhancement: Allow to specify start index for address explorer export and browsing
 - Enhancement: Add `Nunchuk` option to `Export Wallet`
+- Enhancement: Allow to choose only checksum valid last words for seed phrase input  
 - Change: `Passphrase` menu item is no longer offered if BIP39 passphrase
   already in use. Use `Restore Master` with ability to keep or purge current
   passphrase wallet settings.